rename directories names to proper one
